Remove the duplicate resize hack
It turns out this problem also included a bunch of other data corruption
issues. I don't really know how that got in there, but I don't think
this was a data-structure issue. We should revert to an assert to crash
hard. We now have some checks that should give us a warning closer to
when the corruption happens next time.